refactor(charts): give the title one module with apply (spec 22 task 2)
PatchTitle, BuildTitleText, PatchExtendedTitle, BuildExtendedTitle and
the rich-text helpers they own become ChartTitleXml.Apply and
ApplyExtended. Creating the element is the branch Apply takes when there
is none, so the writer and the patcher stop being two implementations of
the same mapping. LiteralText stays shared because an axis title is the
same c:tx block under a different parent.

One golden fixture re-baselined, deliberately. bar-titled now carries
<c:autoTitleDeleted val="0"/>. Only PatchTitle kept that element in step
before; the writer never wrote it, so a new titled chart and a retitled
loaded chart disagreed. Unifying them settles the disagreement in favour
of what Excel itself writes, and the element repeats the schema default,
so the chart renders identically. The new characterization test in
ChartTests pins both paths and lands red on the previous commit.

ChartFormatting drops nine members, 16 entry points to 12.
The other seven golden fixtures are byte-identical.
