+++ XLibur.Tests/Excel/Charts/Golden/bar-titled.xml
@@ -1,1 +1,1 @@
-<?xml version="1.0" encoding="utf-8"?><c:chartSpace xmlns:c="http://schemas.openxmlformats.org/drawingml/2006/chart"><c:chart><c:title><c:tx><c:rich><a:bodyPr xmlns:a="http://schemas.openxmlformats.org/drawingml/2006/main" /><a:lstStyle xmlns:a="http://schemas.openxmlformats.org/drawingml/2006/main" /><a:p xmlns:a="http://schemas.openxmlformats.org/drawingml/2006/main"><a:r><a:rPr lang="en-US" /><a:t>Quarterly</a:t></a:r></a:p></c:rich></c:tx><c:overlay val="0" /></c:title><c:plotArea><c:layout /><c:barChart><c:barDir val="col" /><c:grouping val="clustered" /><c:ser><c:idx val="0" /><c:order val="0" /><c:tx><c:v>Units</c:v></c:tx><c:cat><c:strRef><c:f>Data!$A$1:$A$2</c:f></c:strRef></c:cat><c:val><c:numRef><c:f>Data!$B$1:$B$2</c:f></c:numRef></c:val></c:ser><c:axId val="1" /><c:axId val="2" /></c:barChart><c:catAx><c:axId val="1" /><c:scaling><c:orientation val="minMax" /></c:scaling><c:delete val="0" /><c:axPos val="b" /><c:crossAx val="2" /></c:catAx><c:valAx><c:axId val="2" /><c:scaling><c:orientation val="minMax" /></c:scaling><c:delete val="0" /><c:axPos val="l" /><c:title><c:tx><c:rich><a:bodyPr xmlns:a="http://schemas.openxmlformats.org/drawingml/2006/main" /><a:lstStyle xmlns:a="http://schemas.openxmlformats.org/drawingml/2006/main" /><a:p xmlns:a="http://schemas.openxmlformats.org/drawingml/2006/main"><a:r><a:rPr lang="en-US" /><a:t>Units</a:t></a:r></a:p></c:rich></c:tx><c:overlay val="0" /></c:title><c:crossAx val="1" /></c:valAx></c:plotArea><c:plotVisOnly val="1" /></c:chart></c:chartSpace>
+<?xml version="1.0" encoding="utf-8"?><c:chartSpace xmlns:c="http://schemas.openxmlformats.org/drawingml/2006/chart"><c:chart><c:title><c:tx><c:rich><a:bodyPr xmlns:a="http://schemas.openxmlformats.org/drawingml/2006/main" /><a:lstStyle xmlns:a="http://schemas.openxmlformats.org/drawingml/2006/main" /><a:p xmlns:a="http://schemas.openxmlformats.org/drawingml/2006/main"><a:r><a:rPr lang="en-US" /><a:t>Quarterly</a:t></a:r></a:p></c:rich></c:tx><c:overlay val="0" /></c:title><c:autoTitleDeleted val="0" /><c:plotArea><c:layout /><c:barChart><c:barDir val="col" /><c:grouping val="clustered" /><c:ser><c:idx val="0" /><c:order val="0" /><c:tx><c:v>Units</c:v></c:tx><c:cat><c:strRef><c:f>Data!$A$1:$A$2</c:f></c:strRef></c:cat><c:val><c:numRef><c:f>Data!$B$1:$B$2</c:f></c:numRef></c:val></c:ser><c:axId val="1" /><c:axId val="2" /></c:barChart><c:catAx><c:axId val="1" /><c:scaling><c:orientation val="minMax" /></c:scaling><c:delete val="0" /><c:axPos val="b" /><c:crossAx val="2" /></c:catAx><c:valAx><c:axId val="2" /><c:scaling><c:orientation val="minMax" /></c:scaling><c:delete val="0" /><c:axPos val="l" /><c:title><c:tx><c:rich><a:bodyPr xmlns:a="http://schemas.openxmlformats.org/drawingml/2006/main" /><a:lstStyle xmlns:a="http://schemas.openxmlformats.org/drawingml/2006/main" /><a:p xmlns:a="http://schemas.openxmlformats.org/drawingml/2006/main"><a:r><a:rPr lang="en-US" /><a:t>Units</a:t></a:r></a:p></c:rich></c:tx><c:overlay val="0" /></c:title><c:crossAx val="1" /></c:valAx></c:plotArea><c:plotVisOnly val="1" /></c:chart></c:chartSpace>
